// src/types/buttons.ts
export enum ButtonsTypes {
  Prepend = "Добавить в head",
  Append = "Добавить в tail",
  DeleteHead = "Удалить из head",
  DeleteTail = "Удалить из tail",
  AddByIndex = "Добавить по индексу",
  DeleteByIndex = "Удалить по индексу",
  Add = "Добавить",
  Delete = "Удалить",
  Reset = "Очистить",
  Count = "Рассчитать",
  Reverse = "Рассчитать",
}